Add async operations with streaming responses.
It is needed to implement the AsyncMutateRows (#1217). A preview of where I'm
going with it is available here:
https://github.com/dopiera/google-cloud-cpp/commit/17a36d5301d40c9042d238b8816a0ab5741c03dc

This is implemented by adding a AsyncUnaryStreamRpcFunctor, which
implements the stream lifecycle state machine. In order to achieve that,
enqueued async operations had to start reporting to the completion
queue, whether it should keep them or remove them on a notification
event. AsyncUnaryStreamRpcFunctor indicates to the queue that it should
be removed after the whole stream is finished.

Disposition is handled differently in the streamed version of the
functor than it is in the single-RPC. The assumption I made (which I
will double check in integration tests) is that if a stream is
cancelled, the final gRPC status will hold that information. Due to
that, the callbacks for the streaming API don't use AsyncResult. I'd like to
homogenize the non-streaming API later to resemble this one.