Generalize `AsyncRetryUnaryRpc`.
This is a part of `AsyncMutateRows` implementation (#1217).

BulkMutations have to be retried but unlike most RPCs, only
the mutations which were not confirmed should be resent. That creates a
need for retrying operations other than RPCs, hence this generalization.

The generalization is straight-forward - parameterize submitting
requests by introducing an `Operation` class. It can encapsulate unary
regular RPCs (`AsyncUnaryRpc`) or more complex operations like
`BulkMutator` (in the future).

There are no tests because this is just a refactor and doesn't add any
logic.