Implement wrappers for stat(2). (#1319)
We want to warn users when they are using the wrong file type with
UploadFile(). To do this we need to query the file type. This PR
introduces wrappers for stat(2) that work on both Unix-like systems and
Windows.  The wrappers are modeled after the C++17 `std::filesystem`
functions and types, and thus follow the naming conventions from the
standard.