Better management for the cached packages.
Under AppVeyor we were destroying the full vcpkg cache cache every
time the install script changed.  That often destroyed too much, and
the builds cannot complete with an empty cache (sometimes we are
lucky).  Furthermore, sometimes the vcpkg definition of the cached
packages changes, and the dependency trick did not catch that problem.

I think the new approach, where only the packages in the cache that need
rebuilding are cleared, is probably going to give us better build times.