Add support for detecting if running on GCE.
This implements the recommended approach for detecting whether the
program is running on a Google Compute Engine VM instance. It follows
the approach taken by the grpc library for Linux, but for Windows
machines, it checks the registry rather than launching a subshell and
querying WMI, e.g. `Get-WMIObject -Query "SELECT ... FROM Win32_BIOS"`.
The grpc approaches can be seen in their repository:
https://github.com/grpc/grpc/blob/master/src/core/lib/security/credentials/alts/check_gcp_environment_linux.cc