Rework AsyncRetryOp.
The abstraction of AsyncRetryOp has proven to be bad. This commit
reworks it such that operations accumulate the eventual response from
AsyncRetryOp.

That way, AsyncRetryOp's callback is no longer bound to unary RPCs. This
allows for two simplifications:
- the response type for AsyncRetryOp's callback is inferred from operation
- AsyncRetryOp doesn't need to be aware of AsyncUnaryResult

There is a subtle change here: AsyncRetryOp used to rely on disposition
to decide whether a request has been cancelled. Now it relies on gRPC
status. This should be fine given #1308.

The rework of AsyncRetryOp allows for more straight-forward
implementation of AsyncBulkApply.