Implement bucket lock features in BucketMetadata.
This adds the functionality to parse the bucket lock features, and the
code to prepare messages for `Buckets: patch` and `Buckets: update` that
include bucket lock features. I am not including integration tests yet
because those require changes to the testbench and I would rather send a
smaller PR first. Part of the changes for #1029.