Remove Disposition and AsyncUnaryRpcResult (#1341)
This fixes #1308.

The disposition was only meaningful for timers. It wasn't used in unary
RPCs with streaming responses and for those non-streaming it erroneously
indicated whether an operation got cancelled.

The documentation states that "ok" (obtained from AsyncNext) should
always be true for Finish() operation.

After this change:
- disposition is removed from the codebase,
- if an operation is cancelled, we trust that grpc status indicates that
  and pass it to the user,
- if a timer is cancelled, AsyncTimerResult has a bool to reflect it,
- AsyncUnaryRpcResult is removed too, because it doesn't make much sense
  after the removal of disposition,
- if my understanding of the documentation is wrong or gRPC is buggy and
  AsyncNext does return false for a Finish() operation, we translate it
  into an grpc::UNKNOWN.

The change required a modification of a test which assumed that upon
request cancellation AsyncNext returns ok == false.

A test is added to check the newly added behaviour on Finish() returning
ok == false.