Fix handling of missing MD5 hashes.
When using `Objects: compose` the MD5 hashes are not recomputed by GCS.
This commit fixes the testbench to reproduce that behavior, because we
miss bugs when it is not emulated properly. I also fixed the composite
hash validator to handle the case where one of the branches does not
receive the expected hash value.