Define interface for storage IdempotencyPolicy.
This is part of the work for #714. This PR introduces no changes in
behavior, it just adds a new abstract class to define the idempotency
policy interface, and a trivial implementation that treats all requests
as idempotent.

This is a large PR, but most of the code is trivial, I hope this is
Okay.