Introduce StrictIdempotencyPolicy.
This is part of the work for #714. It introduces a new (partially
implemented) idempotency policy that only labels as idempotent the
operations that are idempotent in all circumstances. In this PR only the
CRUD operations for buckets are implemented, I plan to create one PR per
resource type to keep the PRs reasonably sized. Because the policy is
not fully implemented there are no integration tests yet.