Enable CRC32C checksums by default. (#1344)
With this PR the APIs to insert objects, upload objects, and download
objects compute the CRC32C checksum by default. This is in addition to
the MD5 hash, application developers can disable one or both of these.

When using `Objects: compose` the MD5 hashes are not recomputed by GCS.
This commit fixes the testbench to reproduce that behavior, because we
miss bugs when it is not emulated properly. I also fixed the composite
hash validator to handle the case where one of the branches does not
receive the expected hash value.