Implement the base class for the shared state. (#1351)
Futures and promises keep a shared state that tracks whether the promise
is satisfied, if the value has been retrieved, the exception (if any),
and hold the (type erased) continuation. In this PR we introduce this
base class and unit tests for it, as part of the work for #1345.

The class is incomplete, and there are a lot of support classes around
it that will be introduced in future PRs, as it is, this PR is getting
too large already.