Remove git commit annotation for release builds. (#1353)
With this PR versions are no longer annotated with the GIT version
number.  This makes it easier for package managers to build our
releases, while still allowing us to know what commit we used for a test
or benchmark in development.

When we create releases we will need to commit a change to the release
branch that sets GOOGLE_CLOUD_CPP_IS_RELEASE to "yes".