Implement examples for object holds. (#1346)
This creates examples showing how to set and release event based holds
and temporary holds. Part of the work for #1029.

Ignore non-writable values in path and update. Thanks to @houglum for
pointing out that returning an error is not what the service does when it
receives a non-writable field in update or patch requests.