Initial support for continuations. (#1373)
This is part of the work for #1345. It implements support for
continuations in the shared state. Continuations are implemented as a
type-erased object associated with the source shared state. They hold
both the functor and the destination shared state. When the source
shared state is satisfied the continuation is invoked, they call the
functor with the source shared state input, and store the results (if
any) into the output shared state.

In this PR this is only implemented for `void` as the shared state, so
it is a bit boring, but it is a good start, and more code would make the
PR even larger than it already is.