Adds support for Compute Engine credentials. (#1381)
Adds support for Compute Engine credentials.

GoogleDefaultCredentials() now checks if the program is running on a GCE instance, and if so (and no higher-precedence credentials were already found), the Credentials pointer returned has a derived type of ComputeEngineCredentials.  I've also added support for creating this new credential type directly.

In theory, this credential type should "just work" for App Engine Flexible Environment custom environments as well, but that theory has not been tested yet. Confirmation for GAEF support is still being tracked in #579.