Initial implementation of future<void>::then() (#1391)
This is part of the work for #1345. It implements future<void>::then(),
because future<T> is not implemented yet, it only works for functors
that return `void`.

gcc-4.9 did not handle a protected type alias used in the definition of
a lambda correctly. Creating a local alias for the protected type alias
fixed the problem. Left a comment explaining this too.