Add missing unit tests.
Some of the code was untested, it would be in a future PR, but I think
it is better to add the tests now. That showed some of the code to be
incomplete (and/or broken), which also would have been fixed in a future
PR, but better to have consistent commits.  Sorry about the mess.