Move future<void>::then_impl() to a separate file. (#1414)
The definition of `future<T>::then_impl()` may need to reference
future<void>, and the definition of `future<void>::then_impl()` may
need to reference `future<T>`. To break the cycles we need to move
the definition of this template functions to a separate file, where
`future<T>`, `future<void>` and (soon) `future<R&>` are all visible.
This is part of the work for #1345.