Implement unwrapping .then(). (#1421)
When passing a functor that returns `R` to `.then()` the return value is
`future<R>`. If implemented naively, `.then() would be incovenient when
the return value `R` is actually a `future<R2>`. The naive version would
return `future<future<R2>>` and the application developer would need to
wait for the first future, and then wait for the wrapped `future<R2>`.

The Technical Specification we are following requires `.then()` to
unwrap the future in this case, and return `future<R2>`. This unwrapping
future waits for the (implicit) `future<future<R2>>` to become
satisfied, and then waits for the `future<R2>` to become satisfied.

This is specially important for us, because we expect that most
`bigtable::*::Async*()` member functions will return a `future<R2>`, so
to use them in `.then()` (some lambda adapters) we need to support
unwrapping.

Part of the work for #1345.