Support WithObjectMetadata option in InsertObjectMedia.
When inserting objects it is now possible to set the metadata at the
same time as we upload the media. This is useful when the application
wants to set labels or other object metadata attributes and save the
additional request. The library automatically switches to a multipart
upload when this option is used, but prefers a simple upload otherwise.
Part of the changes for #1311.