Imeplemnt AsyncAwaitConsistency.
This fixes #1256.

The request kicks off a `AsyncGenerateConsistencyToken` and on success
passes it to `AsyncPollCheckConsistency` which in turn calls the user
callback.

Unfortunately, the implementation is not trivial because we have to
handle cancellations. In order for the whole request to be cancellable,
we need to create an `AsyncOperation`, which will cancel either of the
two underlying requests depending on the timing.

This implementation introduces a new request (`AsyncAwaitConsistency`),
which holds all the necessary data to launch the following
`AsyncPollCheckConsistency` after `AsyncGenerateConsistencyToken`
finishes and implements `AsyncOperation` so that it can return itself to
the user as a handle for cancellation.

By adding `CheckConsistencyFunctor` and
`ConsistencyTokenGeneratedFunctor` nested classes I achieved two things:
avoided copying the user-provided functor (as I would have to do with
lambdas) and managed to keep AsyncAwaitConsistency not a template, which
makes its use easier.