Wrappers to capture callback results in promises. (#1453)
In the `noex` layer we use callbacks to report the result of
asynchronous operations.  In the layer with exceptions we will use
future<T>, and the application can block, or attach a continuation.

To convert between the two layers we need something that basically
stores the parameters of the callback into a `promise<T>`, with C++11 we
need a helper class to do this mapping.