Client option to select resumable vs. simple uploads. (#1465)
This is part of the changes for #559. When uploading a file we will
select the type of upload (resumable vs. simple) based on the size of
the file. Small regular files can be loaded to memory and use a simple
retry strategy. Large files, or non-regular files, need to use resumable
uploads and a more sophisticated retry strategy.