Implement AsyncRetryAndPollUnaryRpc.
This fixes #1397.

It glues together retrying an RPC to issue an API call whose result is a
`longrunning.Operation` and then polling until that operation is
complete.

The implementation and tests are heavilly inspired by AsyncAwaitConsistency.

The added tests show how it will be used to implement calls like
AsyncCreateCluster.