Implement AsyncRetryAndPollUnaryRpc. (#1469)
This fixes #1397.

It glues together retrying an RPC to issue an API call whose result is a
`longrunning.Operation` and then polling until that operation is
complete.

The implementation and tests are heavilly inspired by AsyncAwaitConsistency.

The added tests show how it will be used to implement calls like
AsyncCreateCluster.

Move AsyncRetryAndPollUnaryRpc to its own file.

MSVC complained about a tool large file and was probably right.

AsyncRetryAndPollUnaryRpc: use static_assert.

The use of `std::enable_if` is an overkill in this case - SFINAE is
unlikely. `static_assert` makes forward declarations of this class more
concise.  They will be needed for declaring this class as a friend.