Add configuration parameters to install locks.
This fixes #1070. The locking callbacks are not installed if (a) the
application already installed other locking callbacks, or (b) the
application disables them by setting a parameter in the ClientOptions.

The tests are annoying because we can only test one case in each binary,
once set the callbacks stay set forever.