Introduce AsyncLoopOp. (#1539)
`AsyncPollOp` and `AsyncRetryOp` share a non-tirivial amount of code
with each other. Moreover, a class for fetching responses in pages also
would (#1477 has shown that reusing `AsyncPollOp` is ugly).

`AsynLoopOp` is exactly what all of them have in common - running
requests in a loop, potentially sleeping between them and allowing for
cancellations.

There are no test because `AsyncRetryOp` and `AsyncPollOp` are already
tested.