Fix flaky test.
Some of the integration tests for bucket operations were flaky, they
were all operating on the same bucket, so if another integration build
was running at the same time they would not work. This fixes those tests
by creating a bucket (with a random name) for each relevant test.

I also removed a number of pre-conditions in the examples (IfMatchEtag)
that have no effect.

This fixes #1370.