Implement CurlClient function to resume uploads. (#1571)
To fully support resumable uploads we need to be able to, well, resume
an upload. This PR adds functions to restore a resumable upload session
based on the session id. Once created the function queries GCS to
discover what is the latest received byte. In future PRs we will use
this to continue an upload after restoring the session.