Make ccache bigger for coverage builds. (#1567)
We need a bigger cache for the code coverage build. While most builds
take 500MiB, the code coverage build takes nearly 1.5GiB.

When I ran this the build time went from ~37m to ~30m, that is a big gain,
and the coverage build is the longest build so it speeds up the iterations.