WriteObject() uses resumable uploads if requested. (#1597)
With this change the application can select resumable uploads for
streams created via WriteObject().

* Make ObjectStream::Suspend() rvalue-ref qualified.

Because the function is destructive it should be rvalue-ref qualified.
That forces applications to say `std::move(foo).Suspend()` and that
makes it obvious that the `foo` object is unusable. In some cases they
may even get a nice warning from the IDE or compiler.