URL-escape the object names. (#1605)
The object names are not necessarily URL safe, so we need to URL-escape
them.  Also noticed that the entity names may not be
(foo+bar@example.com is a valid email address). I did not URL-escape the
bucket names in this PR because I am not sure we should, they are
supposed to be URL safe, if the application provides an invalid bucket
name this falls under the "do-not-duplicate-server-side-validation" ADR.

This fixes #1590.