Require rvalue-ref for as_proto_move().
The as_proto_move() member functions are destructive, the steal the
contents of the wrapper class. This change limits their use to
rvalue references, so the developer (this is mostly us) notices that
they are destroying the thing (and sometimes get a nice warning).

This is a breaking change, it modifies the public API in a minor way,
but does break existing code. It also breaks the ABI.