Include Copyright notices in the .bzl file.
Per @mbrukman we want to do this for every committed file.

Because the copyright year is supposed to be the year when the file is
created, the CMake functions that generate these files take an optional
YEAR parameter.  If set, that is the copyright year in the generated
file. Right now the default is 2018 because that is when all the current
files were generated. We can change this in the future.