In GCP_LOG() the level should be concatenated without expansion.
This fixes #1623. On macOS certain system headers use DEBUG as a
macro to detect, well, if the code is being built for debugging,
and Bazel obliges (as it should) by defining the macro to 1. If
we expand the macro before concatenation then GCP_LOG(DEBUG) has
a GCP_LS_1 in its expansion, while we want GCP_LS_DEBUG.

See this for additional background:

https://github.com/bazelbuild/bazel/issues/3513