Allow generation of dependency graphs with CMake.
Sometimes it is good to have a dependency graph, either to be able to
explain things, or to debug build problems. CMake can generate them for
us (with the `--graphviz=...` option), but it produces better results
when executed in the CI builds with `TEST_INSTALL=yes`. Add an
environment variable to generate the graphs, I do not think we will use
it much, just from time to time.