Fix macOS+Bazel build on Kokoro.
For some reason the version of Bazel installed on macOS (0.11 when
I checked, but others are reported in the documentation) does not
support the `url` attribute in the `http_archive` rule. All versions
seem to support `urls`, so use that instead. That is not enough
though, because gRPC still uses the `url` attribute.

I have sent a PR to the gRPC folks, and filed a bug to update the
default version of Bazel on Kokoro. Meanwhile, the script manually
updates the Bazel version on each build.

To debug this I removed a lot of unused stuff in the build script, I
like it better as a smaller script so I left those changes too.