For external builds disable ldap.
libcurl automatically discovers available libraries to support
additional protocols, such as LDAP, or SSH. That is hard to use with
ExternalProject_Add() because it requires to maintain the list of
indirect dependencies when using static libraries. With shared objects
it is not a problem (the .so loads anything else it needs).

With this change we disable all protocols that are not interesting in
libcurl. For applications that want to build with a richer libcurl
library they can simply set `GOOGLE_CLOUD_CPP_CURL_PROVIDER` to
`package`.