Prototype nothrow overloads (#1713)
Move the decision to throw or not to `storage::Client` or
`storage::noex::Client`. This makes it possible to chose the error
reporting approach on a per-call basis.

Introduce a CreateNotification() and ListNotifications() overloads
that return StatusOr<T> to signal errors. Implement the existing
overloads in terms of the new function.