Use StatusOr<T> to return errors. (#1723)
Most of the code was using std::pair<Status, T>, but now that we have a
StatusOr template class this is a much better representation. There is
no impact to the public API in this PR, as all functions affected are
either in storage::internal, or they were private functions.