StatusOr<T> when there is no default constructor. (#1725)
When `T` lacks a default constructor we need to use some (more or less
standard) tricks to implement the class. I modeled these changes after
optional<T>, where the expectation is that move-assignments and
move-constructors minimize the number of temporaries.

Thanks to @devbww for the detailed review and all the useful suggestions.