Non throwing object read stream (#1750)
When possible, use `status()` to report errors in `ObjectReadStream`.
Unfortunately the standard library assumes that exceptions are enabled,
and a `std::basic_istreambuf<>` can only report errors by raising
exceptions. When exceptions are disabled this is, hmmm, difficult, so we
must rely on the application dutifully checking the `status()` flag.