Non throwing object write stream (#1749)
Unless the application explicitly requests exceptions by setting the
exception mask. By default `std::ios_base::badbit` is set when an error
is detected, and the ObjectWriteStream saves the status with a more
detailed error message.

This change will reduce code coverage because many errors that used to
be propagated as exceptions now require an extra if(), but we do not
have test infrastructure to produce these error conditions on demand.

Wrote a test to verify that resumable upload errors are reported
correctly. Fixed the testbench to properly validate preconditions on a
resumable upload. Then found a bug in how we handle error status for
resumable upload failures.