Use the right protoc compiler. (#1766)
The rules in CompileProtos.cmake were picking up whatever protoc binary
happened to be in the path. This is "Not Right"[tm], because the version
in the path may not be compatible with the version of the protobuf
libraries found by CMake.

With this change we use the `protobuf::protoc` target. We define this
target for external projects, and for versions of FindProtobuf that do
not provide the target.

This fixes #1762.