Initialize member variable.
This was also reported by Coverity Scan. In this case we are using an
atomic bool to force the right semantics around synchronization, though
we are not interested in the actual value. It is easier to fix the
problem than to explain to Coverity Scan why it is Okay to have an
unitialized variable here.