Separate Docker build output directories. (#1774)
This makes it easier to run multiple different Docker-based builds in a
developer workstation. Each build gets a different directory output
directory based on the settings for the build. Before this change one
had to clear the output directory for (say) a gcc build before building
with clang for the same distribution.

I also removed the `cached-` prefix from the image names, that made
sense at some point when we actually tried to cache the images, but no
longer does.