Automatically upload Coverity Scan results [skip ci].
If COVERITY_SCAN_TOKEN is defined it automatically prepares and uploads
the result tarball to Coverity Scan for analysis. I am not sure we will
ever use this in Travis, the build takes ~45 minutes on my laptop with 8
cores and a local SSD. Travis has a limit of 50 minutes and only 2
cores, it will (almost certainly) run out of time before finishing. But
having a single command to run locally is still useful.