Fix error handling in curl_multi_wait(). (#1796)
This was a defect reported by Coverity Scan. Apparently I forgot how
`return` works, and was still thinking "Call this function which will
throw an exception on error." Sigh. In any case, we were lucky (or
unlucky) in that the bug is just a performance issue: the loop would
spin instead of sleeping between calls to curl_multi_wait().