Change *Notification() functions to return StatusOr. (#1792)
In this change all the `storage::Client::*Notification()` functions are
changed to return `StatusOr<>`. This change also introduces new helpers
to write unit tests for functions that return `StatusOr<>` instead of
throwing exceptions.

The integration tests are changed to check the if the returned
`StatusOr<>` is `.ok()`, and then we use the `operator*()` and
`operator->()` to use the `StatusOr<T>` as a pointer to `T`.

This fixes #1634, #1635, #1636, and #1637.