Implement Client::ListObjects with StatusOr (#1799)
Note that `ListObjects` itself does not return a `StatusOr`, it returns
a ListObjectReader which is always a valid result, even if reading
fails. The change is to actually return `StatusOr<T>` from `operator*()`.

The build for Ubuntu:18.04 breaks with a optional<StatusOr<T>>. It
probably should be fixed, we must be doing something too complicated. On
the other hand, we do not need an `optional<StatusOr<T>>`, a "plain"
`StatusOr<T>` is flexible enough to store a `T` or a "you tried to read
an invalid iterator", so do that instead.