Use installer script for bazel.
Installing via homebrew broke this morning, so switch to just
downloading and running the installation script. This pins the version
of Bazel we will use for the CI builds, which is good (more reliable
builds), and bad (users may use a newer version and we do not know it
breaks).