Fix handling of 308 responses. (#1831)
When resuming an upload that has received no bytes the server can
respond with 308 and exclude the `Range:` header. The client should
interpret that as "oh, you have no committed bytes, cool I will start
from 0 then." We were interpreting that as "The server is broken, you
are supposed to tell me what was the last byte." I guess I just read the
documentation wrong.

I fixed the testbench to behave like the production service does in this
case. That broke one of the integration tests (yay!) and then fixed the
code to handle the response message the "Right Way":tm: