Move exceptions outside of Credentials class ctors. (#1868)
We don't want the constructors of Credentials subclasses to throw
exceptions, so the parsing steps that could have thrown them have been
moved outside of the ctors, which now accept structs containing the
already-parsed information.

The tests have also been cleaned up a bit to test parsing separately
from credential instantiation, and the test names were tweaked to be
more specific (e.g. to indicate if we're testing that something is supposed
to fail, as opposed to a success case). Duplicate tests were also removed
(e.g. one that tested for failure when a field was left out was removed, as
we had another test farther down that removed each field and checked
for failures in each case).

This is part of the work to address #1806. Will send a follow-up PR to
start returning StatusOr instead of throwing.