Re-added "Include Order" section to the style guide (#1876)
When I previously changed the style guide I accidentally removed this section. When I realized that the old style guide for this project differed from the GSG on this point, my intent was to "fix" our code to align it with the GSG's guidance on this point. But then I started doing that it all the includes looked less sensible to me. So after some reflection and discussion w/ other Googler's about this style point, I think I prefer if this project violates the GSG on this point and keeps the current include ordering that we've been using.

Thanks for doing this the right way from the start, Carlos! :-) Sorry I accidentally removed this section.