Disable `make install` for external projects.
When compiling against dependencies obtained via `ExternalProject_Add`
we should disable the `make install` rule: the dependencies may not be
installed in the system, or may not be the same version used to compile
the code. A future PR will improve the documentation to explain how to
build the system with `make install` support.