Create Kokoro configs for more builds. (#1928)
This creates the Kokoro configurations for the
UndefinedBehaviorSanitizer build, the build without exceptions, and the
build for CentOS 7. These are simple builds to move out of Travis, so we
start with them. The first step is to create these configs. Once merged
I will add the internal to Google configuration files, and once I verify
those work I will remove the corresponding builds for the .travis.yml
file.