Allow specifying non-default scopes and subject. (#1987)
Fixes #1866, which mentions that the default scopes for
ServiceAccountCredentials are not always enough. Allows passing an
empty optional for the set of scopes, in which case the default scope is
used, or an optional containing a set, in which case the strings in the set
are concatenated into a comma-delimited string. This change also allows
specifying the subject (for domain-wide delegation purposes) used for
authorization grants.