Stops using grpc's DO_NOT_USE enum (#2013)
As the name implies, we shouldn't be using `grpc::StatusCode::DO_NOT_USE`. Additionally, I'm removing our own `kDoNotUse` status code. We don't need it. If unspecified, an enum class has an underlying type of `int`, so we can test stringifying any value within the range of `int`. In this case, I'm using `42` to make it clearish that I'm choosing an arbitrary value. Of course, if our legit status code values ever cover the value 42, this test will break, but that seems unlikely and the fix should be straightforward.