New scripts to extract README.md instructions.
The Dockerfiles in ci/test-readme/ can be used to generate a big chunk
of the README.md file (specifically the instructions on how to install
development tools). I created a script to extract this portion, and used
it to update the README.md file.  I also added instructions on how to
generate the instructions.