Make InstanceAdmin emulator thread-safe. (#2027)
TIL: gRPC servers are multi-threaded by default:

https://grpc.io/docs/tutorials/basic/c.html#implementing-routeguide

This emulator was *generally* fine because we send one request at a
time. That is, except in the ListAllClusters() member function where we
actually create two threads and send two requests simultaneously. This
(I think) explains the flaky tests.

The good news is that this is only the emulator, the library is fine.