Enable OpenSSL when compiling libcurl as external. (#2029)
We disabled most protocols when compiling libcurl as an external project
because they often require additional libraries and managing that was
becoming difficult. But I got overenthusiastic and I removed OpenSSL
too. With local integration tests that is not a problem (we do not use
HTTPS in that case), but failed with production tests. *And* our
CI builds do not use external projects when running against production.

This fixes the immediate problem. A future PR will need to fix the
testing gap.