Only require googletest if it is needed. (#2035)
When compiling with -DBUILD_TESTING=OFF the tests are disabled. This is
useful for package maintainers who build the same version over and over
and do not want to rebuild the tests every time, nor do they want the
dependency on a library (googletest) that is not needed at run-time.