Increase timeouts for CompletionQueue tests. (#2045)
The CompletionQueue test failed once with a `wait_for(50_ms)` timing
out. Evidently the timeout is too short, so increase it to 500_ms.
Personally I dislike timeout values in unit tests, but this is a test to
verify that timers set via the gRPC completion queue work, so a generous
timeout is fine I think. Note that the test will typically complete much
faster than 500ms.