Increase timeouts in future<> tests. (#2047)
Some of the tests verify that waiting for an event for X milliseconds
gets a response. The timeouts were typically 10ms or 50ms, but that is
too short for the CI systems (we got a failure with the 10ms timeout).

I increased all the actual waits to 500ms. I am not sure that is long
enough, but since I have only seen one failure due to timeouts with 10ms
it cannot be that terrible. Note that this will not slow down the tests,
this the the maximum time to wait, but if the scheduler and thread
communication works as usual the tests complete in less than 1ms.

Some calls to wait_for() or wait_until() are expecting a timeout or
are "waiting" for an event that already took place. In those cases I
left the values unchanged.