Keep Apply callbacks in MutationData.
Currently, `Batch` holds all callbacks in a `deque<>`. This is enough if
all mutations finish at the same time. When we make use of streaming, it
will no longer be true.

When mutations finish, we want to release resources held by them early,
so that we can let more mutations into the `BatchedMutator`. In order to
make it possible, this change starts keeping the callbacks in a hash-map
indexed by indices of `SingleRowMutations` in `BulkMutation`.

We will also need to update stats as individual mutations finish, not
the whole batch, so the callbacks are packed together with the
per-`SingleRowMutation` stats into `MutationData`.