Fix flaky integration test(s) for Bigtable. (#2057)
We used to create a separate table for each test case in the integration
tests. That makes the test case easier to write (and reason about), but
creating a table is very slow in the production tests, and one can
easily hit the daily or peak quota for Bigtable:

https://cloud.google.com/bigtable/quotas#operation-quotas

In this PR we change the integration tests to share a common table. I
had to normalize the table schema: typically we had column families
named "family{1,2,3}" but sometimes it was "fam{0,1,2,3}". I also
changed the SetUp() function to drop all the rows in the table, so at
least the table is clean.

Note that some of the instance admin and table admin tests still create
tables, but those are typically disabled when running against
production.

In addition, the PR fine tunes the message size for the original flaky test.

The intention of the test is to verify that we can write and read cells much
larger than the default gRPC maximum message size (~4MiB).

This default is too small for Cloud Bigtable, where cells can be much
larger than 4MiB. The library achieves this by setting a larger value
(~256MiB) for the maximum message size.

The test was creating a row of ~250MiB. I think testing with 128MiB is
more reasonable. Large enough to prove that we are not using the 4MiB
setting, yet small enough to finish quickly and avoid trouble with any
message overhead.

I was able to run the test with the new setting about 1000 times without
problem. But note that other issues (contention in the bigtable master,
quota limits on the number of tables created by hour / day / etc.) could
also turn this test flaky. That is why I did the refactoring for CreateTable().